修改bug增加订单删除功能

+++ ruoyi-txs/src/main/java/com/ruoyi/txs/service/impl/TxsCustomerServiceImpl.java
@@ -48,6 +48,9 @@ public class TxsCustomerServiceImpl implements TxsCustomerService {
     @Override
     @Transactional
     public int deleteByIds(String ids) {
+        if (StringUtils.isEmpty(ids)) {
+            return 0;
+        }
         Long[] idArr = Convert.toLongArray(ids);
         return dao.deleteByIds(idArr);
     }
